refactor(reporting): extract report engine from compose into modules/reporting
The report-spec compiler/runner and drill-through derivation lived at the
composition root (compose/) though they are self-contained feature logic —
the first slice of factoring compose/ into dedicated modules.

Move (behavior unchanged): report.go, derivation.go, reporthandlers.go,
derivation_test.go -> backend/internal/modules/reporting/. report.go was at
the 500-LOC cap, so its SQL-building half is split into reportsql.go.

Exported surface: Engine, New, Run, RunAdHocPlan, Derive, Request, Aggregate,
Outcome, Handlers, NewHandlers, FieldNotAllowedError. The ad-hoc/datasource
path needs schemaFields (which stays in compose, used by provider too), so it
is injected into New() as a func over neutral datasource types — no import
cycle.

Rewired the three consumers: server.go (HTTP handlers), registry.go (the
agent/MCP report tool), provider.go (datasource ad-hoc plan).

Verified: go build ./... green; go vet clean; reporting unit tests pass;
go-file-length gate green (0 waivers, stale report.go waiver removed);
compose test binaries still compile.
