+++ scripts/go-file-length-waivers.txt
@@ -2,4 +2,3 @@
 # the gate landed (scripts/check-go-file-length.sh). Shrinking is allowed;
 # growing is not. When a file drops to 500 or below, DELETE its line here.
 # Format: <frozen-line-count> <path>
-501 backend/internal/compose/report.go
